Fixed issue that made VocLearningDataSet not appear in solution(? feedback needed)
I added (using "Add...") the two files VocLearningDataSet.Designer.cs and VocLearningDataSet.xsd so they are included in the solution explorer as they were not before (they were just on the same folder). Please give feedback if this does not work on you (if the files are still not visible in the solution explorer).